[BBB-84] 🛠️Refactor: Study, User 관련 Entity 및 JpaRepository Infra로 이동
