--- domain/integrated/src/main/java/com/bombombom/devs/study/Constants.java
@@ -1,13 +1,0 @@
-package com.bombombom.devs.study;
-
-public class Constants {
-
-    public static final int MAX_CAPACITY = 20;
-    public static final int MAX_WEEKS = 52;
-
-    public static final int MAX_RELIABILITY_LIMIT = 100;
-    public static final int MAX_PENALTY = 100_000;
-    public static final int MAX_DIFFICULTY_LEVEL = 29;
-    public static final int MAX_PROBLEM_COUNT = 20;
-
-}
