+++ domain/src/main/java/com/bombombom/devs/domain/study/repository/StudyRepository.java
@@ -1,8 +1,8 @@
 package com.bombombom.devs.domain.study.repository;
 
+import com.bombombom.devs.common.Pageable;
 import com.bombombom.devs.domain.study.model.Page;
 import com.bombombom.devs.domain.study.model.Study;
-import com.bombombom.devs.domain.study.vo.Pageable;
 import java.util.Optional;
 
 public interface StudyRepository {
@@ -10,5 +10,5 @@ public interface StudyRepository {
     Optional<Study> findStudyWithUsersById(Long id);
 
     Page<Study> findAllWithUserAndBook(Pageable pageable);
-    
+
 }
