+++ infra/mysql/build.gradle
@@ -2,4 +2,5 @@ dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
     runtimeOnly 'com.mysql:mysql-connector-j'
     compileOnly project(':domain')
+    implementation project(':common')
 }
