+++ infra/mysql/src/main/java/com/bombombom/devs/mysql/study/entity/AlgorithmProblemAssignmentSolveHistory.java
@@ -1,4 +1,4 @@
-package com.bombombom.devs.study.models;
+package com.bombombom.devs.mysql.study.entity;
 
 import com.bombombom.devs.global.audit.BaseEntity;
 import com.bombombom.devs.user.models.User;
