+++ infra/mysql/src/main/java/com/bombombom/devs/mysql/study/entity/Study.java
@@ -1,9 +1,8 @@
 package com.bombombom.devs.mysql.study.entity;
 
-import com.bombombom.devs.domain.study.StudyStatus;
-import com.bombombom.devs.domain.study.StudyType;
+import com.bombombom.devs.domain.study.enums.StudyStatus;
+import com.bombombom.devs.domain.study.enums.StudyType;
 import com.bombombom.devs.mysql.BaseEntity;
-import com.bombombom.devs.user.models.User;
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.ConstraintMode;
