+++ infra/mysql/src/main/java/com/bombombom/devs/mysql/study/repository/RoundRepository.java
@@ -1,4 +1,4 @@
-package com.bombombom.devs.study.repository;
+package com.bombombom.devs.mysql.study.repository;
 
 import com.bombombom.devs.study.models.Round;
 import java.time.LocalDate;
