+++ infra/mysql/src/main/java/com/bombombom/devs/mysql/study/repository/StudyEntityRepository.java
@@ -1,9 +1,9 @@
 package com.bombombom.devs.mysql.study.repository;
 
+import com.bombombom.devs.common.Pageable;
 import com.bombombom.devs.domain.study.model.Page;
 import com.bombombom.devs.domain.study.model.Study;
 import com.bombombom.devs.domain.study.repository.StudyRepository;
-import com.bombombom.devs.domain.study.vo.Pageable;
 import java.util.Optional;
 import org.springframework.stereotype.Repository;
